+++ core/src/main/java/hudson/security/HudsonPrivateSecurityRealm.java
@@ -75,6 +75,9 @@ import javax.crypto.SecretKeyFactory;
 import javax.crypto.spec.PBEKeySpec;
 import jenkins.model.Jenkins;
 import jenkins.security.FIPS140;
+import jenkins.security.NonePasswordComplexityRule;
+import jenkins.security.PasswordComplexityException;
+import jenkins.security.PasswordComplexityRule;
 import jenkins.security.SecurityListener;
 import jenkins.security.seed.UserSeedProperty;
 import jenkins.util.SystemProperties;
