+++ core/src/main/resources/hudson/security/Messages.properties
@@ -45,14 +45,6 @@ HudsonPrivateSecurityRealm.CreateAccount.UserNameInvalidCharacters=User name mus
 HudsonPrivateSecurityRealm.CreateAccount.UserNameInvalidCharactersCustom=User name must match the following expression: {0}
 HudsonPrivateSecurityRealm.CreateAccount.InvalidEmailAddress=Invalid e-mail address
 HudsonPrivateSecurityRealm.CreateAccount.UserNameAlreadyTaken=User name is already taken
-HudsonPrivateSecurityRealm.CreateAccount.PasswordTooShort=Password must be at least {0} characters long.
-HudsonPrivateSecurityRealm.CreateAccount.PasswordRequiresUppercase=Password must contain at least one uppercase letter (A-Z).
-HudsonPrivateSecurityRealm.CreateAccount.PasswordRequiresLowercase=Password must contain at least one lowercase letter (a-z).
-HudsonPrivateSecurityRealm.CreateAccount.PasswordRequiresDigit=Password must contain at least one digit (0-9).
-HudsonPrivateSecurityRealm.CreateAccount.PasswordRequiresSpecialCharacter=Password must contain at least one special character.
-
-BasicPasswordComplexityRule.DisplayName=Basic password complexity
-NonePasswordComplexityRule.DisplayName=None
 
 FullControlOnceLoggedInAuthorizationStrategy.DisplayName=Logged-in users can do anything
