+++ core/src/main/resources/jenkins/security/Messages.properties
@@ -70,3 +70,11 @@ ResourceDomainConfiguration.Invalid=Not a valid URL.
 ResourceDomainConfiguration.SameAsJenkinsRoot=Cannot use the same host name for both Jenkins URL and resource root URL.
 ResourceDomainConfiguration.SameAsCurrent=You are currently accessing Jenkins through a URL similar to the proposed resource root URL. Saving this URL might remove your access to Jenkins.
 
+BasicPasswordComplexityRule.DisplayName=Basic password complexity
+BasicPasswordComplexityRule.PasswordTooShort=Password must be at least {0} characters long.
+BasicPasswordComplexityRule.PasswordRequiresUppercase=Password must contain at least one uppercase letter (A-Z).
+BasicPasswordComplexityRule.PasswordRequiresLowercase=Password must contain at least one lowercase letter (a-z).
+BasicPasswordComplexityRule.PasswordRequiresDigit=Password must contain at least one digit (0-9).
+BasicPasswordComplexityRule.PasswordRequiresSpecialCharacter=Password must contain at least one special character (e.g. !@#$%^&*).
+NonePasswordComplexityRule.DisplayName=None
+
